### 2021-05-09 #3268 @whedon
[PRE REVIEW]: neworder: a dynamic microsimulation framework for python
**Submitting author:** @virgesmith (<a href="http://orcid.org/0000-0002-9951-6642">Andrew Smith</a>)
**Repository:** <a href="https://github.com/virgesmith/neworder" target ="_blank">https://github.com/virgesmith/neworder</a>
**Version:** 0.3.0
