+++ paper/paper.md
@@ -14,7 +14,7 @@ authors:
 affiliations:
  - name: School of Law, University of Leeds, UK
    index: 1
-   date: 7 May 2021
+date: 8 May 2021
 bibliography: paper.bib
 ---
 
