+++ paper/paper.bib
@@ -1,23 +1,38 @@
 
-@misc{noauthor_liam2_nodate,
-	title = {{LIAM2} website},
-	url = {https://liam2.plan.be/pages/about.html},
-	urldate = {2021-05-08},
+@article{lewis_simulation_1979,
+	title = {Simulation of nonhomogeneous poisson processes by thinning},
+	volume = {26},
+	copyright = {Copyright © 1979 Wiley Periodicals, Inc., A Wiley Company},
+	issn = {1931-9193},
+	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/nav.3800260304},
+	doi = {10.1002/nav.3800260304},
+	abstract = {A simple and relatively efficient method for simulating one-dimensional and two-dimensional nonhomogeneous Poisson processes is presented The method is applicable for any rate function and is based on controlled deletion of points in a Poisson process whose rate function dominates the given rate function In its simplest implementation, the method obviates the need for numerical integration of the rate function, for ordering of points, and for generation of Poisson variates.},
+	language = {en},
+	number = {3},
+	urldate = {2021-05-07},
+	journal = {Naval Research Logistics Quarterly},
+	author = {Lewis, P. a. W. and Shedler, G. S.},
+	year = {1979},
+	pages = {403--413},
 }
 
-@misc{noauthor_m-o-p-dpolice-supply-demand_2021,
+@misc{birks_m-o-p-dpolice-supply-demand_2021,
 	title = {M-{O}-{P}-{D}/{Police}-{Supply}-{Demand}},
-	copyright = {MIT License
-        , 
-      
-        MIT License},
+	copyright = {MIT License},
 	url = {https://github.com/M-O-P-D/Police-Supply-Demand},
 	abstract = {Repository for Supply ABM for Police Supply and Demand Models},
 	urldate = {2021-05-08},
 	publisher = {M.O.P.D.},
+	author = {Birks, Dan and Smith, Andrew P},
 	month = may,
 	year = {2021},
 	note = {original-date: 2021-03-16T08:56:14Z},
+}
+
+@misc{noauthor_liam2_nodate,
+	title = {{LIAM2} website},
+	url = {https://liam2.plan.be/pages/about.html},
+	urldate = {2021-05-08},
 }
 
 @misc{smith_neworder_2021,
@@ -75,21 +90,4 @@
 	author = {Government of Canada, Statistics Canada},
 	month = sep,
 	year = {2009},
-}
-
-@article{lewis_simulation_1979,
-	title = {Simulation of nonhomogeneous poisson processes by thinning},
-	volume = {26},
-	copyright = {Copyright © 1979 Wiley Periodicals, Inc., A Wiley Company},
-	issn = {1931-9193},
-	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/nav.3800260304},
-	doi = {https://doi.org/10.1002/nav.3800260304},
-	abstract = {A simple and relatively efficient method for simulating one-dimensional and two-dimensional nonhomogeneous Poisson processes is presented The method is applicable for any rate function and is based on controlled deletion of points in a Poisson process whose rate function dominates the given rate function In its simplest implementation, the method obviates the need for numerical integration of the rate function, for ordering of points, and for generation of Poisson variates.},
-	language = {en},
-	number = {3},
-	urldate = {2021-05-07},
-	journal = {Naval Research Logistics Quarterly},
-	author = {Lewis, P. a. W. and Shedler, G. S.},
-	year = {1979},
-	pages = {403--413},
 }
