+++ paper/paper.bib
@@ -1,93 +1,91 @@
 
-@article{lewis_simulation_1979,
-	title = {Simulation of nonhomogeneous poisson processes by thinning},
-	volume = {26},
-	copyright = {Copyright © 1979 Wiley Periodicals, Inc., A Wiley Company},
-	issn = {1931-9193},
-	url = {https://onlinelibrary.wiley.com/doi/abs/10.1002/nav.3800260304},
-	doi = {10.1002/nav.3800260304},
-	abstract = {A simple and relatively efficient method for simulating one-dimensional and two-dimensional nonhomogeneous Poisson processes is presented The method is applicable for any rate function and is based on controlled deletion of points in a Poisson process whose rate function dominates the given rate function In its simplest implementation, the method obviates the need for numerical integration of the rate function, for ordering of points, and for generation of Poisson variates.},
-	language = {en},
-	number = {3},
-	urldate = {2021-05-07},
-	journal = {Naval Research Logistics Quarterly},
-	author = {Lewis, P. a. W. and Shedler, G. S.},
-	year = {1979},
-	pages = {403--413},
-}
-
-@misc{birks_m-o-p-dpolice-supply-demand_2021,
-	title = {M-{O}-{P}-{D}/{Police}-{Supply}-{Demand}},
-	copyright = {MIT License},
+@software{birks_m-o-p-dpolice-supply-demand_2021,
+	title = {M-O-P-D/Police-Supply-Demand},
+	rights = {{MIT} License         ,                 {MIT} License},
 	url = {https://github.com/M-O-P-D/Police-Supply-Demand},
-	abstract = {Repository for Supply ABM for Police Supply and Demand Models},
-	urldate = {2021-05-08},
+	abstract = {Repository for Supply {ABM} for Police Supply and Demand Models},
 	publisher = {M.O.P.D.},
 	author = {Birks, Dan and Smith, Andrew P},
-	month = may,
-	year = {2021},
+	urldate = {2021-05-08},
+	date = {2021-05-07},
 	note = {original-date: 2021-03-16T08:56:14Z},
 }
 
-@misc{noauthor_liam2_nodate,
-	title = {{LIAM2} website},
+@online{noauthor_liam2_nodate,
+	title = {{LIAM}2 website},
 	url = {https://liam2.plan.be/pages/about.html},
 	urldate = {2021-05-08},
 }
 
-@misc{smith_neworder_2021,
+@online{smith_neworder_2021,
 	title = {neworder},
 	url = {https://neworder.readthedocs.io/en/stable/},
-	urldate = {2021-05-07},
-	journal = {neworder documentation},
+	titleaddon = {neworder documentation},
 	author = {Smith, Andrew P},
-	year = {2021},
+	urldate = {2021-05-07},
+	date = {2021},
 }
 
 @book{belanger_microsimulation_2017,
-	series = {The {Springer} {Series} on {Demographic} {Methods} and {Population} {Analysis}},
-	title = {Microsimulation and {Population} {Dynamics}: {An} {Introduction} to {Modgen} 12},
-	isbn = {9783319446622},
-	shorttitle = {Microsimulation and {Population} {Dynamics}},
+	title = {Microsimulation and Population Dynamics: An Introduction to Modgen 12},
+	isbn = {978-3-319-44662-2},
 	url = {https://www.springer.com/gp/book/9783319446622},
+	series = {The Springer Series on Demographic Methods and Population Analysis},
+	shorttitle = {Microsimulation and Population Dynamics},
[…]
