+++ paper/paper.bib
@@ -1,91 +1,95 @@
 
-@software{birks_m-o-p-dpolice-supply-demand_2021,
-	title = {M-O-P-D/Police-Supply-Demand},
-	rights = {{MIT} License         ,                 {MIT} License},
-	url = {https://github.com/M-O-P-D/Police-Supply-Demand},
-	abstract = {Repository for Supply {ABM} for Police Supply and Demand Models},
-	publisher = {M.O.P.D.},
-	author = {Birks, Dan and Smith, Andrew P},
-	urldate = {2021-05-08},
-	date = {2021-05-07},
-	note = {original-date: 2021-03-16T08:56:14Z},
-}
-
-@online{noauthor_liam2_nodate,
-	title = {{LIAM}2 website},
+@misc{noauthor_liam2_nodate,
+	title = {{LIAM2} website},
 	url = {https://liam2.plan.be/pages/about.html},
 	urldate = {2021-05-08},
 }
 
-@online{smith_neworder_2021,
+@misc{noauthor_m-o-p-dpolice-supply-demand_2021,
+	title = {M-{O}-{P}-{D}/{Police}-{Supply}-{Demand}},
+	copyright = {MIT License
+        ,
+
+        MIT License},
+	url = {https://github.com/M-O-P-D/Police-Supply-Demand},
+	abstract = {Repository for Supply ABM for Police Supply and Demand Models},
+	urldate = {2021-05-08},
+	publisher = {M.O.P.D.},
+	month = may,
+	year = {2021},
+	note = {original-date: 2021-03-16T08:56:14Z},
+}
+
+@misc{smith_neworder_2021,
 	title = {neworder},
 	url = {https://neworder.readthedocs.io/en/stable/},
-	titleaddon = {neworder documentation},
-	author = {Smith, Andrew P},
 	urldate = {2021-05-07},
-	date = {2021},
+	journal = {neworder documentation},
+	author = {Smith, Andrew P},
+	year = {2021},
 }
 
 @book{belanger_microsimulation_2017,
-	title = {Microsimulation and Population Dynamics: An Introduction to Modgen 12},
-	isbn = {978-3-319-44662-2},
+	series = {The {Springer} {Series} on {Demographic} {Methods} and {Population} {Analysis}},
+	title = {Microsimulation and {Population} {Dynamics}: {An} {Introduction} to {Modgen} 12},
+	isbn = {9783319446622},
+	shorttitle = {Microsimulation and {Population} {Dynamics}},
 	url = {https://www.springer.com/gp/book/9783319446622},
-	series = {The Springer Series on Demographic Methods and Population Analysis},
-	shorttitle = {Microsimulation and Population Dynamics},
 	abstract = {This book is a practical, step-by-step introduction to microsimulation in demography. It shows how to use Modgen, a powerful and free microsimulation platform built by Statistics Canada. The authors' hands-on explanation of model development will help readers make their own. The book teaches how to create and run a simple cohort model with a single fixed-rate event, and builds upon this concept. It introduces how to develop both a single state life table as well as a multiple increment-decrement life table using the tools provided by Modgen. The authors illustrate how to easily upgrade an existing model by adding new modules and new dimensions as determinants of a risk already modeled. The integration of a fertility module and a base population allows the user to bring new actors into the simulation and transform a cohort-based model into a population-based one. The final addition of an international migration module allows the user to accomplish fully open, multi-regional projections. This accessible introduction will be of interest to researchers and students in population studies and other social sciences. It will also appeal to anyone interested in the computational modeling of population dynamics.},
+	language = {en},
+	urldate = {2021-05-07},
[…]
