+++ neworder/__init__.py
@@ -1,2 +1,5 @@
 __version__ = "0.3.0"
+
 from _neworder_core import *
+from . import domain
+
