+++ neworder/__init__.py
@@ -1,5 +1,5 @@
 __version__ = "0.3.0"
 
 from _neworder_core import *
-from . import domain
+from .domain import *
 
