+++ examples-requirements.txt
@@ -26,5 +26,5 @@ pytz==2020.1
 requests==2.22.0
 retrying==1.3.3
 six==1.14.0
-urllib3==1.25.8
+urllib3==1.26.5
 webencodings==0.5.1
