+++ Dockerfile
@@ -11,10 +11,11 @@ RUN apt-get update -y \
  && rm -rf /var/lib/apt/lists/*
 
 # install python deps
+RUN pip install -U pip pytest
 RUN python -m pip install -r examples-requirements.txt
 
 # use setup.py to get pytest (not in requirements.txt)
-RUN python setup.py install && python setup.py test && mpiexec -n 2 python setup.py pytest
+RUN pip install -e . && python -m pytest && mpiexec -n 2 python -m pytest
 
 ENV DISPLAY :0
 
