+++ appveyor.yml
@@ -3,35 +3,35 @@
 environment:
   matrix:
 
-    - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
-      PYTHON: "C:\\Python37-x64"
-      PYTHON_ARCH: 64
-      MSVC_VERSION: "Visual Studio 19 Win64"
+    # - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
+    #   PYTHON: "C:\\Python37-x64"
+    #   PYTHON_ARCH: 64
+    #   MSVC_VERSION: "Visual Studio 19 Win64"
 
-    - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
-      PYTHON: "C:\\Python38-x64"
-      PYTHON_ARCH: 64
-      MSVC_VERSION: "Visual Studio 19 Win64"
+    # - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
+    #   PYTHON: "C:\\Python38-x64"
+    #   PYTHON_ARCH: 64
+    #   MSVC_VERSION: "Visual Studio 19 Win64"
 
-    - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
-      PYTHON: "C:\\Python39-x64"
-      PYTHON_ARCH: 64
-      MSVC_VERSION: "Visual Studio 19 Win64"
+    # - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
+    #   PYTHON: "C:\\Python39-x64"
+    #   PYTHON_ARCH: 64
+    #   MSVC_VERSION: "Visual Studio 19 Win64"
 
     # - APPVEYOR_BUILD_WORKER_IMAGE: Visual Studio 2019
     #   MINICONDA: "C:\\Miniconda38-x64"
 
-    - APPVEYOR_BUILD_WORKER_IMAGE: Ubuntu
-      PYTHON: "3.7"
-      APPVEYOR_YML_DISABLE_PS_LINUX: true
+    # - APPVEYOR_BUILD_WORKER_IMAGE: Ubuntu
+    #   PYTHON: "3.7"
+    #   APPVEYOR_YML_DISABLE_PS_LINUX: true
 
-    - APPVEYOR_BUILD_WORKER_IMAGE: Ubuntu2004
-      PYTHON: "3.8"
-      APPVEYOR_YML_DISABLE_PS_LINUX: true
+    # - APPVEYOR_BUILD_WORKER_IMAGE: Ubuntu2004
+    #   PYTHON: "3.8"
+    #   APPVEYOR_YML_DISABLE_PS_LINUX: true
 
-    - APPVEYOR_BUILD_WORKER_IMAGE: Ubuntu2004
-      PYTHON: "3.9"
-      APPVEYOR_YML_DISABLE_PS_LINUX: true
+    # - APPVEYOR_BUILD_WORKER_IMAGE: Ubuntu2004
+    #   PYTHON: "3.9"
+    #   APPVEYOR_YML_DISABLE_PS_LINUX: true
 
     - APPVEYOR_BUILD_WORKER_IMAGE: Ubuntu2004
       MINICONDA: true
@@ -43,17 +43,17 @@ environment:
       PYTHON: "3.8"
       APPVEYOR_YML_DISABLE_PS_LINUX: true
 
-    - APPVEYOR_BUILD_WORKER_IMAGE: macOS
-      PYTHON: "3.7"
-      APPVEYOR_YML_DISABLE_PS_LINUX: true
+    # - APPVEYOR_BUILD_WORKER_IMAGE: macOS
[…]
