+++ .github/workflows/coverage.yml
@@ -1,0 +1,38 @@
+# This workflow will install Python dependencies, run tests and lint with a variety of Python versions
+# For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
+
+name: Python coverage
+
+on:
+  push:
+    branches: [ master ]
+  pull_request:
+    branches: [ master ]
+
+jobs:
+  build:
+
+    runs-on: ubuntu-latest
+    strategy:
+      fail-fast: false
+      matrix:
+        python-version: [ 3.8 ]
+    steps:
+    - uses: actions/checkout@v2
+    - name: "pip: Python ${{ matrix.python-version }} coverage"
+      uses: actions/setup-python@v2
+      with:
+        python-version: ${{ matrix.python-version }}
+    - name: Install dependencies
+      run: |
+        python -m pip install --upgrade pip
+        python -m pip install pybind11 pytest
+    - name: Build
+      run: |
+        CFLAGS=-coverage python -m pip install -e .
+    - name: Test
+      run: |
+        python -m pytest
+    - name: Upload
+      run: |
+        bash <(curl -s https://codecov.io/bash) -Z
