### 2021-06-07 #3268 @kthyng
@virgesmith I can't quite tell what your submission is about in order to suggest an editor (to me, the words describing the work are generic and could be from many different fields). Do you have more editor suggestions? Also I see you describe the work as in pre-release — this can be a sign that the work is not ready to be published by JOSS. When do you plan to release version 1.0?
