### 2021-06-07 #3268 @virgesmith
@kthyng it is rather generic, I'm afraid - the techniques* are commonly applied in social sciences, e.g. geography, demography, epidemiology, I've personally used it for demographic modelling and more recently criminology. But it can be applied to any application where modelling the evolution and interactions of individual entities is required.

Re: editors I tried to pick social-scientists with C++ and python skills, but this combination is relatively rare so don't really have any other suggestions, sorry

Regarding "pre-release" I should remove this now. It is a stable, tested, packaged and documented product. That is not to say that I have no plans to extend it. At the risk of complicating my response to your first question I intend to enhance its agent-based modelling capabilities in the medium term - there does seem to be a "gap in the market" for python-based ABM tools.

&ast; the terms microsimulation, agent-based modelling, individual-based modelling overlap hugely and are often used interchangably
