### 2021-06-09 #3268 @platipodium
> At the risk of complicating my response to your first question I intend to enhance its agent-based modelling capabilities in the medium term - there does seem to be a "gap in the market" for python-based ABM tools.

I certainly agree, there definitely is a gap in the market for python-based ABM tools.
