+++ neworder/domain.py
@@ -1,16 +1,22 @@
 from os import stat
 import numpy as np
 
+def _bounce(point, min, max):
+  for d in range(len(point)):
+    if point[d] < min[d]: point[d] = min[d] - point[d]
+    if point[d] > max[d]: point[d] = max[d] - point[d]
+  return point
+
 class Domain:
 
   # Edge behaviour
+  UNBOUNDED = 0
   WRAP = 1
   CONSTRAIN = 2
-  #BOUNCE = 3
+  BOUNCE = 3
 
   def __init__(self, dim, edge, continuous):
     self.__dim = dim
-    assert edge in [Domain.WRAP, Domain.CONSTRAIN]
     self.__edge = edge
     self.__continuous = continuous
 
@@ -27,16 +33,20 @@ class Domain:
     return self.__continuous
 
 class Space(Domain):
+  """ Continuous rectangular n-dimensional domain """
 
-  @classmethod
-  def unbounded(cls, dim):
-    assert dim 
-    return cls(np.full(dim, -np.inf), np.full(dim, +np.inf))
+  @staticmethod
+  def unbounded(dim):
+    assert dim
+    return Space(np.full(dim, -np.inf), np.full(dim, +np.inf), edge=Domain.UNBOUNDED)
 
-  def __init__(self, min, max, edge = Domain.CONSTRAIN):
+  def __init__(self, min, max, edge=Domain.CONSTRAIN):
     assert len(min) and len(min) == len(max)
-    super().__init__(len(min), edge, True)
-    #self.dim = len(min)
+    super().__init__(len(min), edge, edge)
+
+    # Space supports all edge behaviours
+    assert edge in [Domain.UNBOUNDED, Domain.WRAP, Domain.CONSTRAIN, Domain.BOUNCE]
+
     self.min = min
     self.max = max
 
@@ -44,44 +54,83 @@ class Space(Domain):
   def extent(self):
     return self.min, self.max
 
-  def move(self, points, delta, ungroup=False):
+  def move(self, positions, velocities, delta_t, ungroup=False):
+    """ Returns translated positions AND velocities """
     # group tuples into a single array if necessary
-    if type(points) == tuple:
-      points = np.column_stack(points)
-    if type(delta) == tuple:
-      delta = np.column_stack(delta)
+    if type(positions) == tuple:
+      positions = np.column_stack(positions)
+    if type(velocities) == tuple:
+      velocities = np.column_stack(velocities)
 
-    assert points.dtype == float
-    assert delta.dtype == float
-    assert points.shape[-1] == self.dim and delta.shape[-1] == self.dim
-    if self.edge == Domain.CONSTRAIN:
-      result = np.clip(points + delta, self.min, self.max)
+    assert positions.dtype == float
+    assert velocities.dtype == float
+    assert positions.shape[-1] == self.dim and velocities.shape[-1] == self.dim
+    if self.edge == Domain.UNBOUNDED: 
+      p = positions + velocities * delta_t
+      v = velocities
+    elif self.edge == Domain.CONSTRAIN: 
+      p = positions + velocities * delta_t
+      v = velocities
+      hitmin = np.where(p < self.min, 1, 0)
+      p = np.where(hitmin, self.min, p)
+      v = np.where(hitmin, 0, v)
+      hitmax = np.where(p > self.max, 1, 0)
+      p = np.where(hitmax, self.max, p)
+      v = np.where(hitmax, 0, v)   
+    elif self.edge == Domain.BOUNCE:
+      p = positions + velocities * delta_t
+      v = velocities
+      hitmin = np.where(p < self.min, 1, 0)
+      p = np.where(hitmin, 2*self.min - p, p)
+      v = np.where(hitmin, -v, v)
[…]
