### 2021-06-09 #3268 @platipodium
@arfon How can I update my reviewer profile in  this [list of people ](https://bit.ly/joss-reviewers) ?
