### 2021-06-09 #3268 @danielskatz
> > INVALID DOIs
> >
> > * https://doi.org/10.1002/nav.3800260304 is INVALID because of 'https://doi.org/' prefix
>
> Hey @whedon, I guess you forgot to update your doi rules, as the reference with https://doi.org/ is now the recommended way to cite DOis

But this isn't how bibtex works - leave off the prefix in the bibtex entry, and it will be added back in the pandoc/tex processing
