### 2021-06-09 #3268 @zeroset
> wave @zeroset, @platipodium, @tresoldi - Would a couple of you be willing to review this submission to JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Hi @danielskatz, unfortunately i already have another ongoing review and too less time. Sorry.
