+++ README.md
@@ -10,7 +10,7 @@
 [![DOI](https://zenodo.org/badge/111997710.svg)](https://zenodo.org/badge/latestdoi/111997710)
 
 <!--[![Build Status](https://travis-ci.org/virgesmith/neworder.png?branch=master)](https://travis-ci.org/virgesmith/neworder)-->
-[![python (pip) build](https://github.com/virgesmith/humanleague/actions/workflows/pip-package.yml/badge.svg)]()
+[![python (pip) build](https://github.com/virgesmith/neworder/actions/workflows/pip-package.yml/badge.svg)]()
 [![Build status](https://ci.appveyor.com/api/projects/status/oycn4is2insoiun7?svg=true)](https://ci.appveyor.com/project/virgesmith/neworder)
 [![codecov](https://codecov.io/gh/virgesmith/neworder/branch/master/graph/badge.svg?token=g5mDOcjGTD)](https://codecov.io/gh/virgesmith/neworder)
 [![Documentation Status](https://readthedocs.org/projects/neworder/badge/?version=latest)](https://neworder.readthedocs.io/en/latest/?badge=latest)
@@ -40,8 +40,6 @@
 but the module will work perfectly well in serial mode without these.
 
 ## Installation
-
-Note: *neworder* is currently in pre-release. Potentially breaking changes to the API cannot be ruled out before a full 1.0 release.
 
 ### PyPI
 
