+++ untested/boids/run_model.py
@@ -3,7 +3,7 @@ import neworder as no
 import matplotlib.pyplot as plt
 from boids2d import Boids
 
-N = 10
+N = 100
 
 m = Boids(N)
 
