### 2021-06-10 #3268 @tresoldi
@danielskatz I have checked the code and the draft. The library seems robust and useful (I will probably consider it for a project we are just beginning to plan!) and I would be happy to take this review.

Note, however, that an ABM tool lies a bit too much outside my domains of expertise (historical linguistics and evolutionary studies in general, especially phylogenetics). While I am confident I can can review it as research software (usability, documentation, etc., Python & C++ code), we could probably find people better suited to judge its novelty, the state of the field, references, etc.
