### 2021-06-10 #3268 @platipodium
@tresoldi

> Note, however, that an ABM tool lies a bit too much outside my domains of expertise (historical linguistics and evolutionary studies in general, especially phylogenetics). While I am confident I can can review it as research software (usability, documentation, etc., Python & C++ code), we could probably find people better suited to judge its novelty, the state of the field, references, etc.

I am proficient in ABM modeling and applications, so we should make a good complementary reviewer team!
