### 2021-06-10 #3351 @whedon
[REVIEW]: neworder: a dynamic microsimulation framework for python
**Submitting author:** @virgesmith (<a href="http://orcid.org/0000-0002-9951-6642">Andrew Smith</a>)
**Repository:** <a href="https://github.com/virgesmith/neworder" target ="_blank">https://github.com/virgesmith/neworder</a>
**Version:** v1.0.0
