### 2021-06-14 virgesmith/neworder#60 @platipodium
Uses deprecated pandas accessor: "Passing list-likes to .loc or [] with any missing labels is no longer supported."
```
File "/opt/local/Library/Frameworks/Python.framework/Versions/3.9/lib/python3.9/site-packages/pandas/core/indexing.py", line 1321, in _validate_read_indexer
    raise KeyError(
KeyError: "Passing list-likes to .loc or [] with any missing labels is no longer supported. The following labels were missing: Int64Index([6719, 5046, 2701, 1495, 1034,\n            ...\n            7209, 2425, 8597, 2434, 1847],\n           dtype='int64', name='id', length=221). See https://pandas.pydata.org/pandas-docs/stable/user_guide/indexing.html#deprecate-loc-reindex-listlike"
```
