### 2021-06-14 virgesmith/neworder#60 @danielskatz
This is associated with openjournals/joss-reviews#3351
