### 2021-06-14 virgesmith/neworder#61 @platipodium
Multiple redundant calculation in parallel running of serial examples
When issueing, e.g.,

`mpirun -np 4 python examples/people/model.py`

the independent processors duplicate work:

```
[py 1/4] check OK: time=2051-01-01 size=329717 mean_age=41.77, pct_female=49.30 net_migration=1130.0 (20175-19045.0)
[py 3/4] check OK: time=2051-01-01 size=329717 mean_age=41.77, pct_female=49.30 net_migration=1130.0 (20175-19045.0)
[py 2/4] check OK: time=2051-01-01 size=329717 mean_age=41.77, pct_female=49.30 net_migration=1130.0 (20175-19045.0)
```

Is this intentional or is it just wasting CPU cycles?
