### 2021-06-14 #3351 @platipodium
Created https://github.com/virgesmith/neworder/issues/61 for duplicated CPU work
