### 2021-06-14 virgesmith/neworder#62 @platipodium
 Wolf-sheep example default **should not** issue warnings
On running the wolf-sheep model, the default installation **should not** issue warnings, but it does:

python examples/wolf-sheep/model.py

```
/neworder/examples/wolf-sheep/wolf_sheep.py:257: UserWarning: Attempting to set identical bottom == top == 0 results in singular transformations; automatically expanding.
  self.ax_t3.set_ylim([0, max(n/max(len(self.wolves), 1))])
/neworder/examples/wolf-sheep/wolf_sheep.py:260: RuntimeWarning: invalid value encountered in true_divide
  for rect, h in zip(self.ax_ss, n/len(self.sheep)):
/neworder/examples/wolf-sheep/wolf_sheep.py:262: UserWarning: Attempting to set identical bottom == top == 0 results in singular transformations; automatically expanding.
  self.ax_t4.set_ylim([0, max(n/max(len(self.sheep), 1))])
```
