### 2021-06-14 #3351 @platipodium
And another one on warnings in default installation https://github.com/virgesmith/neworder/issues/62.  I guess I should learn how to link to the upstream repository....
