### 2021-06-14 virgesmith/neworder#63 @platipodium
ModuleNotFoundError: No module named 'neworder.domain'
When running

`mpirun -np 4 python examples/n-body/model.py`

```
  File "/neworder/examples/n-body/n_body.py", line 5, in <module>
    from neworder.domain import Space
ModuleNotFoundError: No module named 'neworder.domain'
    from n_body import NBody
```
