### 2021-06-14 #3351 @platipodium
> @platipodium - see the answer in [#3351 (comment)](https://github.com/openjournals/joss-reviews/issues/3351#issuecomment-858564082)

I guess my attempt to mention failed :=( See  https://github.com/virgesmith/neworder/issues/63#issuecomment-860665099
