### 2021-06-14 virgesmith/neworder#63 @platipodium
This is associated with [openjournals/joss-reviews#3351](https://github.com/openjournals/joss-reviews/issues/3351)
