### 2021-06-14 virgesmith/neworder#63 @virgesmith
@platipodium can you confirm you're using it looks like you are using the latest *release* of the package with the bleeding edge version of the examples, which won't necessarily work (the domain stuff is a work-in-progress).

You have 2 options: use the released package with the versioned examples from the website (e.g. download link on https://neworder.readthedocs.io/en/stable/examples/n-body/), or build the latest version of the master branch in the repo - see  https://neworder.readthedocs.io/en/stable/developer/ and use the example code in the repo
