### 2021-06-14 virgesmith/neworder#63 @platipodium
Well-spotted: using release package and bleeding edge examples...  I'll report back on this
