### 2021-06-14 virgesmith/neworder#61 @virgesmith
This particular example isn't set up to be (usefully) run in parallel - if you execute via `mpiexec` it will just do identical runs in each process, because the seeding strategy is set to `neworder.MonteCarlo.deterministic_identical_stream`. You can change the seeding strategy if you want different random streams for each process - if you haven't already seen it i'd recommend reading this page https://neworder.readthedocs.io/en/stable/tips/
