### 2021-06-14 a1172ae
fix issue where speed could evolve to an invalid value (<=0) which causes a population explosion. this also stops the warning messages that were occurring #62
