### 2021-06-14 virgesmith/neworder#60 @virgesmith
this example is another that isn't intended to be run in parallel - it doesn't have *any* logic to divide up the computation to run on multiple processes. having said that, the actual error is distinctly unobvious (and I don't get this error running serially) so I'm just going to add an assertion that it's running in serial mode, hope this is sufficient
