### 2021-06-14 b3d53b3
ensure wolf-sheep is run in serial mode
