+++ src/Module.cpp
@@ -131,8 +131,9 @@ PYBIND11_MODULE(_neworder_core, m)
     .def(py::init([](no::LinearTimeline& t, const py::function& s) { return no::Model(std::make_unique<no::LinearTimeline>(t), s); }), model_init_lineartimeline_docstr,"timeline"_a, "seeder"_a)
     .def(py::init([](no::NumericTimeline& t, const py::function& s) { return no::Model(std::make_unique<no::NumericTimeline>(t), s); }), model_init_numerictimeline_docstr,"timeline"_a, "seeder"_a)
     .def(py::init([](no::CalendarTimeline& t, const py::function& s) { return no::Model(std::make_unique<no::CalendarTimeline>(t), s); }), model_init_calendartimelime_docstr,"timeline"_a, "seeder"_a)
-    .def("timeline", &no::Model::timeline, py::return_value_policy::reference, model_timeline_docstr)
-    .def("mc", &no::Model::mc, py::return_value_policy::reference, model_mc_docstr)
+    // properties are readonly only in the sense you can't assign to them; you CAN call their mutable methods
+    .def_property_readonly("timeline", &no::Model::timeline, model_timeline_docstr)
+    .def_property_readonly("mc", &no::Model::mc, model_mc_docstr)
     .def("modify", &no::Model::modify, model_modify_docstr, "r"_a)
     .def("step", &no::Model::step, model_step_docstr)
     .def("check", &no::Model::check, model_check_docstr)
