+++ examples/hello-world/model.py
@@ -76,7 +76,7 @@ class HelloWorld(neworder.Model):
     Returns: NoneType
     """
     # randomly make some people talkative
-    self.population.talkative = self.mc().hazard(self.p_talk, len(self.population)).astype(bool)
+    self.population.talkative = self.mc.hazard(self.p_talk, len(self.population)).astype(bool)
   # !step!
 
   # !finalise!
