+++ docs/examples/src.md
@@ -1,6 +1,6 @@
 !!! note "Examples Source Code"
     Source code for all the examples can be downloaded using one of the the links below:
 
-    **[neworder-0.3.0-examples-src.tgz](./neworder-0.3.0-examples-src.tgz)** | **[neworder-0.3.0-examples-src.zip](./neworder-0.3.0-examples-src.zip)**
+    **[neworder-1.0.0-examples-src.tgz](./neworder-1.0.0-examples-src.tgz)** | **[neworder-1.0.0-examples-src.zip](./neworder-1.0.0-examples-src.zip)**
 
     The contained file `requirements.txt` lists the package dependencies.

[…]
