+++ docs/api.md
@@ -103,6 +103,13 @@ Returns the time of the current step in the timeline
 
 ---
 
+## ![class](https://img.shields.io/badge/-class-darkgreen) `Domain`
+
+
+Base class for spatial domains
+
+---
+
 ## ![class](https://img.shields.io/badge/-class-darkgreen) `LinearTimeline`
 
 
@@ -287,15 +294,12 @@ or input is required from an upstream model. The model can be subsequently resum
 For trapping exceptional/error conditions, prefer to raise an exception, or return False from the Model.check() function
 
 
-### ![instance method](https://img.shields.io/badge/-instance method-orange) `mc`
+---
 
-```python
-mc(self: neworder.Model) -> no::MonteCarlo
-```
+### ![property](https://img.shields.io/badge/-property-lightgreen) `mc`
 
 
-Returns the models Monte-Carlo engine
-
+The model's Monte-Carlo engine
 
 ### ![instance method](https://img.shields.io/badge/-instance method-orange) `modify`
 
@@ -321,15 +325,12 @@ Default behaviour raises NotImplementedError.
 This function should not be called directly, it is used by the Model.run() function
 
 
-### ![instance method](https://img.shields.io/badge/-instance method-orange) `timeline`
+---
 
-```python
-timeline(self: neworder.Model) -> no::Timeline
-```
+### ![property](https://img.shields.io/badge/-property-lightgreen) `timeline`
 
 
-Returns the model's timeline object
-
+The model's timeline object
 
 ---
 
@@ -337,233 +338,6 @@ Returns the model's timeline object
 
 
 The model's Monte-Carlo engine with configurable options for parallel execution
-
-### ![instance method](https://img.shields.io/badge/-instance method-orange) `arrivals`
-
-```python
-arrivals(self: neworder.MonteCarlo, lambda: numpy.ndarray[numpy.float64], dt: float, n: int, mingap: float) -> numpy.ndarray[numpy.float64]
-```
-
-
-Returns an array of n arrays of multiple arrival times from a nonhomogeneous Poisson process (with hazard rate lambda[i], time interval dt),
-with a minimum separation between events of mingap. Sampling uses the Lewis-Shedler "thinning" algorithm
-The final value of lambda must be zero, and thus arrivals don't always occur, indicated by a value of neworder.time.never()
-The inner dimension of the returned 2d array is governed by the the maximum number of arrivals sampled, and will thus vary
-
-
-### ![instance method](https://img.shields.io/badge/-instance method-orange) `counts`
-
-```python
-counts(self: neworder.MonteCarlo, lambda: numpy.ndarray[numpy.float64], dt: float) -> numpy.ndarray[numpy.int64]
-```
-
-
-Returns an array of simulated arrival counts (within time dt) for each intensity in lambda
-
-
-### ![function](https://img.shields.io/badge/-function-red) `deterministic_identical_stream`
-
-```python
-deterministic_identical_stream(r: int) -> int
-```
-
-
-Returns a deterministic seed (19937). Input argument is ignored
-
-
-### ![function](https://img.shields.io/badge/-function-red) `deterministic_independent_stream`
-
-```python
-deterministic_independent_stream(r: int) -> int
-```
-
-
-Returns a deterministic seed that is a function of the input (19937+r).
-The model uses the MPI rank as the input argument, allowing for differently seeded streams in each process
-
-
-### ![instance method](https://img.shields.io/badge/-instance method-orange) `first_arrival`
-
-```python
-first_arrival(*args, **kwargs)
-```
-Overloaded function.
-
-```python
[…]
