### 2021-06-28 facf92f
refactor model constrution wrt polymorphic timeline
