### 2021-06-29 virgesmith/neworder#65 @platipodium
Add legal info for contributions
There is no information how contributed code is handled.  Please clarify in your "Contributing" statement, how copyright is retained or transferred upon contributing to `neworder`
