### 2021-06-29 virgesmith/neworder#66 @platipodium
Update copyright year 
Please include 2021 in your copyright year range @openjournals/joss-reviews#3351
