### 2021-06-29 #3351 @platipodium
I am happy with the software and paper pending the two issues with contributing https://github.com/virgesmith/neworder/issues/65 and copyright update https://github.com/virgesmith/neworder/issues/66
