### 2021-06-29 be66c31
update licence. closes #66
