+++ docs/licence.md
@@ -1,6 +1,6 @@
 # MIT License
 
-Copyright &copy; 2017-2020 Andrew P Smith
+Copyright &copy; 2017-2021 Andrew P Smith
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
