### 2021-06-29 virgesmith/neworder#66 @danielskatz
ref: https://github.com/openjournals/joss-reviews/issues/3351
