### 2021-06-29 ec7073b
cherry-pick domain bits
