+++ docs/api.md
@@ -310,6 +310,233 @@ The model's timeline object
 
 The model's Monte-Carlo engine with configurable options for parallel execution
 
+### ![instance method](https://img.shields.io/badge/-instance method-orange) `arrivals`
+
+```python
+arrivals(self: neworder.MonteCarlo, lambda: numpy.ndarray[numpy.float64], dt: float, n: int, mingap: float) -> numpy.ndarray[numpy.float64]
+```
+
+
+Returns an array of n arrays of multiple arrival times from a nonhomogeneous Poisson process (with hazard rate lambda[i], time interval dt),
+with a minimum separation between events of mingap. Sampling uses the Lewis-Shedler "thinning" algorithm
+The final value of lambda must be zero, and thus arrivals don't always occur, indicated by a value of neworder.time.never()
+The inner dimension of the returned 2d array is governed by the the maximum number of arrivals sampled, and will thus vary
+
+
+### ![instance method](https://img.shields.io/badge/-instance method-orange) `counts`
+
+```python
+counts(self: neworder.MonteCarlo, lambda: numpy.ndarray[numpy.float64], dt: float) -> numpy.ndarray[numpy.int64]
+```
+
+
+Returns an array of simulated arrival counts (within time dt) for each intensity in lambda
+
+
+### ![function](https://img.shields.io/badge/-function-red) `deterministic_identical_stream`
+
+```python
+deterministic_identical_stream(r: int) -> int
+```
+
+
+Returns a deterministic seed (19937). Input argument is ignored
+
+
+### ![function](https://img.shields.io/badge/-function-red) `deterministic_independent_stream`
+
+```python
+deterministic_independent_stream(r: int) -> int
+```
+
+
+Returns a deterministic seed that is a function of the input (19937+r).
+The model uses the MPI rank as the input argument, allowing for differently seeded streams in each process
+
+
+### ![instance method](https://img.shields.io/badge/-instance method-orange) `first_arrival`
+
+```python
+first_arrival(*args, **kwargs)
+```
+Overloaded function.
+
+```python
+ first_arrival(self: neworder.MonteCarlo, lambda: numpy.ndarray[numpy.float64], dt: float, n: int, minval: float) -> numpy.ndarray[numpy.float64]
+```
+
+
+Returns an array of length n of first arrival times from a nonhomogeneous Poisson process (with hazard rate lambda[i], time interval dt),
+with a minimum start time of minval. Sampling uses the Lewis-Shedler "thinning" algorithm
+If the final value of lambda is zero, no arrival is indicated by a value of neworder.time.never()
+
+
+```python
+ first_arrival(self: neworder.MonteCarlo, lambda: numpy.ndarray[numpy.float64], dt: float, n: int) -> numpy.ndarray[numpy.float64]
+```
+
+
+Returns an array of length n of first arrival times from a nonhomogeneous Poisson process (with hazard rate lambda[i], time interval dt),
+with no minimum start time. Sampling uses the Lewis-Shedler "thinning" algorithm
+If the final value of lambda is zero, no arrival is indicated by a value of neworder.time.never()
+
+
+### ![instance method](https://img.shields.io/badge/-instance method-orange) `hazard`
+
+```python
+hazard(*args, **kwargs)
+```
+Overloaded function.
+
+```python
+ hazard(self: neworder.MonteCarlo, p: float, n: int) -> numpy.ndarray[numpy.float64]
+```
+
+
+Returns an array of ones (with hazard rate lambda) or zeros of length n
+
+
+```python
[…]
