### 2021-07-01 virgesmith/neworder#65 @virgesmith
TBH I'm not sure what I need to add here, Do you know the legalese? along the lines of it's MIT and contributors retain copyright? or is that in itself enought?
