### 2021-07-01 virgesmith/neworder#65 @platipodium
You need to make sure how to handle contributions to your software by people other than yourself.  E.g.  if I would contribute to `neworder` I would legally be an additional copyright holder to your software.  That in itself is (with MIT license and other FOSS) not a problem.

But it becomes a problem once you have contributions from several people (and therefore many copyright holders) and you would like to change the legal terms of your software, e.g., switching to Apache or GPL license or give some company an exclusive license under special terms.   To do this, you would require a unanimous consent by *all* copyright holders.

To avoid this situation, it is customary to either transfer copyright (which many FOSS developers might dislike) or to retain copyright (which makes it difficult for the maintainer as stated above), or to seek a middle way that is, e.g., explained in detail on http://harmonyagreements.org/index.html
