### 2021-07-02 4c63d7d
add legals to contributing section #65 and remove duplicate copy of licence
