### 2021-07-02 virgesmith/neworder#65 @virgesmith
...now an extra "Legal" section here: https://neworder.readthedocs.io/en/latest/developer/
