### 2021-07-02 4534e31
tweak docs to make contributing section more explicit #65
