+++ untested/boids/boids2d.py
@@ -1,15 +1,12 @@
 import numpy as np
 import pandas as pd
 import neworder as no
-
+from time import sleep
 import matplotlib.pyplot as plt
-
-PAUSE=1e-6
 
 TEST_MODE=False#True
 
 TWO_PI = 2 * np.pi
-
 
 class Boids(no.Model):
 
@@ -25,29 +22,30 @@ class Boids(no.Model):
     self.domain = no.Space(np.array([0,0]), np.array([self.range, self.range]), edge=no.Domain.WRAP)
 
     self.vision = 0.2
-    self.exclusion = 0.01
+    self.exclusion = 0.02
 
     self.sep_max_turn = TWO_PI / 240 # 1.5 degrees
     self.align_max_turn = TWO_PI / 72 # 5 degrees
     self.cohere_max_turn = TWO_PI / 120 # 3 degrees
 
     if TEST_MODE:
-      self.N = 2
+      self.N = 10
       self.boids = pd.DataFrame(
         index = pd.Index(name="id", data=no.df.unique_index(self.N)),
         data={
-          "x": [0.01,1],
-          "y": [0,1],
-          "theta": [np.pi/4, 5*np.pi/4]
+          "x": np.linspace(0, 0.9, 10),
+          "y": np.full(10, 0.5),
+          "theta": np.full(10, np.pi/2)
         })
+      no.log(self.boids)
     else:
-      # initially in [0,1]^3
+      # initially in [0,1]^dim
       self.boids = pd.DataFrame(
         index = pd.Index(name="id", data=no.df.unique_index(self.N)),
         data={
-          "x": self.mc().ustream(N) * self.range,
-          "y": self.mc().ustream(N) * self.range,
-          "theta": self.mc().ustream(N) * TWO_PI # zero is +ve x axis
+          "x": self.mc.ustream(N) * self.range/2,
+          "y": self.mc.ustream(N) * self.range/2,
+          "theta": self.mc.ustream(N) * TWO_PI # zero is +ve x axis
         }
       )
 
@@ -69,17 +67,17 @@ class Boids(no.Model):
     mask = np.repeat(1.0-np.sum(nearest_in_range, axis=0), self.N).reshape(self.N, self.N)
 
     in_range = np.logical_and(np.where(d2 < self.vision**2, 1.0, 0.0), mask).astype(float)
-    self.__cohere(in_range, dx, dy, self.cohere_max_turn)
+    #in_range = np.where(d2 < self.vision**2, 1.0, 0.0).astype(float)
+    #self.__cohere(in_range, dx, dy, self.cohere_max_turn)
     self.__align(in_range, self.align_max_turn)
 
-    (self.boids.x, self.boids.y), _ = self.domain.move((self.boids.x, self.boids.y),
-                                                       (np.cos(self.boids.theta) * self.speed, np.sin(self.boids.theta) * self.speed),
-                                                       self.timeline().dt(), ungroup=True)
-    # for a contrained domain convert updated velocity to angle
-    # self.boids.theta = np.arctan2(vy, vx)
+    (self.boids.x, self.boids.y), (vx, vy) = self.domain.move((self.boids.x, self.boids.y),
+                                                              (np.cos(self.boids.theta) * self.speed, np.sin(self.boids.theta) * self.speed),
+                                                              self.timeline.dt(), ungroup=True)
 
     #self.__track()
     self.__update_visualisation()
+    #sleep(1)
     #self.halt()
 
   def __align(self, in_range, max_turn):
@@ -87,12 +85,15 @@ class Boids(no.Model):
     weights = 1.0 / np.sum(in_range, axis=0)
[…]
