### 2021-07-06 e941695
add boids and update examples
