### 2021-07-08 #3351 @virgesmith
FYI once I've got and addressed the comments from @tresoldi I plan to create a 1.0.0 release to coincide with the JOSS publication...
