### 2021-07-09 #3351 @tresoldi
First of all, sorry for taking longer than promised. I played with neworder for some time, really like it and will recommend the publication. I had even started to model something for my field (diffusion of lexical innovations) that I'd submit as an example, but it is better not to tie the review to it.

I have opened a couple of issues that should be easy to address, related to installation. Overall, the work shows scholarly effort, the package is easy to install, functional, has good performance and good documentation.

I have some issues with the paper in terms of its structure, as it stands apart from most papers published in JOSS and ends up looking a bit tailored for a specialist audience; some general references to microsimulations would be welcome. As for JOSS requested information, most of it can be found, but it could be improved. I'd add a separate "Installation, Examples and Usage" section, possibly taking contents currently under "Statement of need", informing that (a) the library is installed as any normal Python package (this is mentioned in passing, but it would be nice to stress it) and that (b) while examples and usages are too long for a short papers, they are available in the official documentation (which would be better be linked than used as bibliographic reference). Note that these are more of suggestions, and I am happy to accept the work as it is if the author prefers.

Two issues that do need to be addressed are:

  - in the paper, some uncapitalized proper names ("python", "docker") and lowercase acronym ("yaml")
  - in the repository, there are no contribution guidelines; the defaults by GitHub could be used

Once more, sorry for the delay, I promise to do my best to answer within 24 hours.
