### 2021-07-09 #3351 @danielskatz
can you mention openjournals/joss-reviews#3351 in the new issues so that we can track them from here?
