+++ docs/examples/src.md
@@ -3,4 +3,8 @@
 
     **[neworder-1.0.0-examples-src.tgz](./neworder-1.0.0-examples-src.tgz)** | **[neworder-1.0.0-examples-src.zip](./neworder-1.0.0-examples-src.zip)**
 
-    The contained file `requirements.txt` lists the package dependencies.
+    The contained file `requirements.txt` lists the package dependencies of the examples, which can be installed using e.g.: 
+
+    ```sh
+    pip install -r requirements.txt
+    ```
