### 2021-07-09 94bef87
docs and dependencies changes raised in #67, #68
