+++ examples/competing/visualise.py
@@ -22,7 +22,7 @@ def plot_age(model):
   plt.legend(["1st birth", "2nd birth", "3rd birth", "4th birth", "5th birth", "Death"])
   plt.xlabel("Age (y)")
   plt.ylabel("Frequency")
-  plt.savefig("./docs/examples/img/competing_hist_100k.png", dpi=80)
+  #plt.savefig("./docs/examples/img/competing_hist_100k.png", dpi=80)
 
 def plot_parity(model):
   bins = np.arange(model.population.parity.max())-0.25
@@ -30,4 +30,4 @@ def plot_parity(model):
   plt.title("Births during lifetime")
   plt.xlabel("Number of children")
   plt.ylabel("Frequency")
-  plt.savefig("./docs/examples/img/competing_births_100k.png", dpi=80)
+  #plt.savefig("./docs/examples/img/competing_births_100k.png", dpi=80)
