+++ examples/requirements.txt
@@ -1,34 +1,13 @@
-appdirs
-CacheControl
-certifi
-chardet
-colorama
-contextlib2
-cycler
-distlib
-distro
-html5lib
-idna
-ipaddr
-kiwisolver
-lockfile
-matplotlib
-mpi4py
-msgpack
-neworder
-numpy
-packaging
-pandas
-pep517
-Pillow
-progress
-pyparsing
-python-dateutil
-pytoml
-pytz
-requests
-retrying
-scipy
-six
-urllib3
-webencodings
+cycler==0.10.0
+kiwisolver==1.3.1
+matplotlib==3.4.2
+mpi4py==3.0.3
+neworder==1.0.0
+numpy==1.21.0
+pandas==1.3.0
+Pillow==8.3.1
+pyparsing==2.4.7
+python-dateutil==2.8.1
+pytz==2021.1
+scipy==1.7.0
+six==1.16.0
