+++ requirements-developer.txt
@@ -1,29 +1,38 @@
-appdirs>=1.4.3
-CacheControl>=0.12.6
-certifi>=2019.11.28
-chardet>=3.0.4
-colorama>=0.4.3
-contextlib2>=0.6.0
-distlib>=0.3.0
-distro>=1.4.0
-html5lib>=1.0.1
-idna>=2.8
-ipaddr>=2.2.0
-lockfile>=0.12.2
-msgpack>=0.6.2
-numpy>=1.19.5
-packaging>=20.3
-pandas>=1.1.5
-pep517>=0.8.2
-progress>=1.5
-pybind11>=2.6.1
-pyparsing>=2.4.6
-pytest>=6.1.2
+attrs>=21.2.0
+certifi>=2021.5.30
+chardet>=4.0.0
+click>=8.0.1
+ghp-import>=2.0.1
+idna>=2.10
+importlib-metadata>=4.6.1
+iniconfig>=1.1.1
+Jinja2>=3.0.1
+Markdown>=3.3.4
+MarkupSafe>=2.0.1
+mergedeep>=1.3.4
+mkdocs>=1.2.1
+mkdocs-macros-plugin>=0.5.12
+mkdocs-material>=7.1.9
+mkdocs-material-extensions>=1.0.1
+mpi4py>=3.0.3
+numpy>=1.21.0
+packaging>=21.0
+pandas>=1.3.0
+pluggy>=0.13.1
+py>=1.10.0
+pybind11>=2.6.2
+Pygments>=2.9.0
+pymdown-extensions>=8.2
+pyparsing>=2.4.7
+pytest>=6.2.4
 python-dateutil>=2.8.1
-pytoml>=0.1.21
-pytz>=2020.4
-requests>=2.25.0
-retrying>=1.3.3
-six>=1.14.0
-urllib3>=1.26.5
-webencodings>=0.5.1
+pytz>=2021.1
+PyYAML>=5.4.1
+pyyaml-env-tag>=0.1
+requests>=2.25.1
+six>=1.16.0
+termcolor>=1.1.0
+toml>=0.10.2
+urllib3>=1.26.6
+watchdog>=2.1.3
+zipp>=3.5.0
