+++ docs-requirements.txt
@@ -1,7 +1,0 @@
-mkdocs>=1.1.2
-mkdocs-material>=7.1.4
-mkdocs-macros-plugin>=0.5.5
-mkdocs-material-extensions>=1.0.1
-#mkdocstrings==0.13.2
-#Pygments>=2.9.0
-requests
