### 2021-07-10 f3f560b
address review comments re: paper and contribution guidelines https://github.com/openjournals/joss-reviews/issues/3351
