+++ docs/requirements.txt
@@ -1,0 +1,6 @@
+# this file is used by readthedocs.io
+mkdocs>=1.1.2
+mkdocs-material>=7.1.4
+mkdocs-macros-plugin>=0.5.5
+mkdocs-material-extensions>=1.0.1
+requests
