### 2021-07-10 #3351 @virgesmith
@tresoldi you can now see the contributing guidelines on the latest docs build : https://neworder.readthedocs.io/en/latest/ pls confirm you're happy.

also, latest version of paper with proper noun fixes is here https://github.com/virgesmith/neworder/suites/3205032474/artifacts/74112696
