### 2021-07-10 #3351 @virgesmith
@danielskatz @platipodium @tresoldi thanks for your efforts, really appreciated esp in current circumstances
I will get on with the v1.0.0 release now...
