+++ Dockerfile
@@ -12,7 +12,7 @@ RUN apt-get update -y \
 
 # install python deps
 RUN pip install -U pip pytest
-RUN python -m pip install -r examples-requirements.txt
+RUN python -m pip install -r examples/requirements.txt
 
 # use setup.py to get pytest (not in requirements.txt)
 RUN pip install -e . && python -m pytest && mpiexec -n 2 python -m pytest
