### 2021-07-10 #3351 @virgesmith
@danielskatz. 1.0.0 now released, DOI: 10.5281/zenodo.5090120, however you now appear as an author in the metadata
