### 2021-07-11 #3351 @danielskatz
Maybe you can remove me as an author - the metadata is just what zenodo suggests, but you can change it
